requirements.txt
albumentations
hydra-core
omegaconf
rich
torch
torchsummary
torchvision
wandb